requirements.txt
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.6
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
python-decouple==3.4
scp==0.13.3
six==1.15.0
